+++ src/ll/xist/ns/html.py
@@ -196,19 +196,6 @@ class title(xsc.Element):
 	class Attrs(i18nattrs):
 		class id(xsc.IDAttr): pass
 
-	def unwrapHTML(self, node, converter):
-		if isinstance(node, xsc.Element) and node.xmlns==xmlns: # is this one of our own elements => filter it out
-			if isinstance(node, img):
-				node = node["alt"]
-			else:
-				node = node.content.mapped(self.unwrapHTML, converter)
-		return node
-
-	def convert(self, converter):
-		content = self.content.convert(converter)
-		content = content.mapped(self.unwrapHTML, converter)
-		return title(content, self.attrs)
-
 
 class base(xsc.Element):
 	"""